corrección cargado de archivos para participantes

+++ registro_resultados/models.py
@@ -80,7 +80,7 @@ class Participante(models.Model):
     club = models.CharField(max_length=100, verbose_name='club de Registro', null=True, blank=True)
     fecha_nacimiento = models.DateField(null=True, blank=True)
     estatura = models.PositiveIntegerField(verbose_name='estatura (cm)', null=True, blank=True)
-    peso = models.PositiveIntegerField(verbose_name='peso (kg)', null=True, blank=True)
+    peso = models.DecimalField(max_digits=3, decimal_places=2, verbose_name='peso (kg)', null=True, blank=True)
 
     posicion = models.IntegerField(default=0, verbose_name="posición")
     metros = models.DecimalField(default=0, null=True, max_digits=3, decimal_places=2, help_text='En metros')
